CampusDailyAutoSign_src/BusinessCentralLayer/sentinel/alkaid/pyproject.toml
[tool.poetry]
name = "alkaid"
version = "0.1.0"
description = "alkaid"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
nonebot2 = "^2.0.0.a1"
nb-cli = "^0.1.0"

[tool.poetry.dev-dependencies]
nonebot-test = "^0.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
